requirements.txt
streamlit
pypdf
chromadb
pyyaml
google-generativeai
python-dotenv
